requirements.txt
imaplib2>=3.6
python-dotenv
rich>=10.0.0
python-dateutil>=2.8.2
email-validator
flask>=2.0.0
requests>=2.25.0
pdfplumber>=0.10.0
pdfservices-sdk>=4.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
gunicorn>=21.0.0
flask-compress>=1.14
